+++ CliFx/Internal/Polyfills.cs
@@ -9,14 +9,25 @@ namespace System
 
     internal static class Extensions
     {
-        public static bool Contains(this string str, char c) =>
-            str.Any(i => i == c);
+        public static bool Contains(this string str, char c)
+        {
+            return str.Any(i => i == c);
+        }
 
-        public static bool StartsWith(this string str, char c) =>
-            str.Length > 0 && str[0] == c;
+        public static bool StartsWith(this string str, char c)
+        {
+            return str.Length > 0 && str[0] == c;
+        }
 
-        public static bool EndsWith(this string str, char c) =>
-            str.Length > 0 && str[str.Length - 1] == c;
+        public static bool EndsWith(this string str, char c)
+        {
+            return str.Length > 0 && str[str.Length - 1] == c;
+        }
+
+        public static string[] Split(this string str, char separator, StringSplitOptions options = StringSplitOptions.None)
+        {
+            return str.Split(new char[] { separator }, options);
+        }
     }
 }
 
@@ -30,8 +41,21 @@ namespace System.Collections.Generic
             value = pair.Value;
         }
 
-        public static TValue GetValueOrDefault<TKey, TValue>(this IReadOnlyDictionary<TKey, TValue> dic, TKey key) =>
-            dic.TryGetValue(key, out var result) ? result! : default!;
+        public static TValue GetValueOrDefault<TKey, TValue>(this IReadOnlyDictionary<TKey, TValue> dic, TKey key)
+        {
+            return dic.TryGetValue(key, out var result) ? result! : default!;
+        }
+
+        public static bool TryAdd<TKey, TValue>(this IDictionary<TKey, TValue> dic, TKey key, TValue value)
+        {
+            if (!dic.ContainsKey(key))
+            {
+                dic.Add(key, value);
+                return true;
+            }
+
+            return false;
+        }
     }
 }
 
@@ -41,11 +65,15 @@ namespace System.Linq
 
     internal static class Extensions
     {
-        public static HashSet<T> ToHashSet<T>(this IEnumerable<T> source) =>
-            new HashSet<T>(source);
+        public static HashSet<T> ToHashSet<T>(this IEnumerable<T> source)
+        {
+            return new HashSet<T>(source);
+        }
 
-        public static HashSet<T> ToHashSet<T>(this IEnumerable<T> source, IEqualityComparer<T> comparer) =>
-            new HashSet<T>(source, comparer);
+        public static HashSet<T> ToHashSet<T>(this IEnumerable<T> source, IEqualityComparer<T> comparer)
+        {
+            return new HashSet<T>(source, comparer);
+        }
     }
 }
 #endif